+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.4.7] - 2025-05-09
+
+### Fixed
+
+- fix: update error handling in Firestore client tests to return structured JSON error messages
+- fix: enhance Firestore client error handling and ensure proper initialization of Firebase admin instance
+
 ## [1.4.6] - 2025-05-09
 
 ### Fixed